+++ CHANGELOG.rst
@@ -1,3 +1,81 @@
+phase-currentcost (v0.9.14) stable; urgency=low
+-----------------------------------------------
+
+* Correct badge error
+
+Pierre Leray <pierreleray64@gmail.com>  2014-04-25 01:10:00
+
+phase-currentcost (v0.9.13) stable; urgency=low
+-----------------------------------------------
+
+* Fix README problem on Pypi.
+* Change usage to add "sudo" command
+
+Pierre Leray <pierreleray64@gmail.com>  2014-04-25 00:50:00
+
+phase-currentcost (v0.9.12) stable; urgency=low
+-----------------------------------------------
+
+* Try to fix README problem on Pypi.
+
+Pierre Leray <pierreleray64@gmail.com>  2014-04-25 00:45:00
+
+phase-currentcost (v0.9.11) stable; urgency=low
+-----------------------------------------------
+
+* Try to fix README problem on Pypi.
+
+Pierre Leray <pierreleray64@gmail.com>  2014-04-25 00:40:00
+
+phase-currentcost (v0.9.10) stable; urgency=low
+-----------------------------------------------
+
+* PYPI installation is working.
+
+Pierre Leray <pierreleray64@gmail.com>  2014-04-25 00:30:00
+
+phase-currentcost (v0.9.9) stable; urgency=low
+----------------------------------------------
+
+* Fix installation problem on PYPI.
+
+Pierre Leray <pierreleray64@gmail.com>  2014-04-25 00:20:00
+
+phase-currentcost (v0.9.8) stable; urgency=low
+----------------------------------------------
+
+* Modify installation to perform custom installation
+* Add two log.conf file, for dev and prod purpose.
+* Test dev mode.
+* News: add parameter to target log.conf path for test (by default it should be /opt/phase/phase-currentcost.conf)
+* News: add parameter to create currentcost.log for test (by default it should be /var/log/phase/phase-currentcost.log)
+* News: Update test file accordingly
+* Modify installation to perform custom installation following this advice http://stackoverflow.com/questions/15853058/run-custom-task-when-call-pip-install
+* Fix installation bug
+* Add production test in behave and launch it only on travis
+* Improve Readme.md
+* Fix travis dependencies problem
+* Improve todo
+* Upload succesfully package on PYPI
+
+Pierre Leray <pierreleray64@gmail.com>  2014-04-24 23:20:00
+
+phase-currentcost (v0.9.7) stable; urgency=low
+----------------------------------------------
+
+* Working on installation and log configuration
+
+Pierre Leray <pierreleray64@gmail.com>  2014-04-24 22:20:00
+
+phase-currentcost (v0.9.6) stable; urgency=low
+----------------------------------------------
+
+* Change currentcost folder into phase-currencost
+* Complete dosctrings
+* List all message sent to RabbitMQ
+
+Pierre Leray <pierreleray64@gmail.com>  2014-04-18 14:30:00
+
 phase-currentcost (v0.9.5) stable; urgency=low
 ----------------------------------------------